+++ test_runner/src/main/kotlin/ftl/args/ValidateAndroidArgs.kt
@@ -19,6 +19,7 @@ fun AndroidArgs.validate() {
     assertRoboTest()
     assertDirectoriesToPull()
     assertMaxTestShardsByDeviceType()
+    assertParametersConflict()
 }
 
 private fun AndroidArgs.assertAdditionalAppTestApks() {
@@ -98,4 +99,9 @@ private fun AndroidArgs.throwMaxTestShardsLimitExceeded(): Nothing {
     throw FlankConfigurationError(
         "max-test-shards must be >= ${IArgs.AVAILABLE_VIRTUAL_SHARD_COUNT_RANGE.first} and <= ${IArgs.AVAILABLE_VIRTUAL_SHARD_COUNT_RANGE.last} for virtual devices, for physical devices max-test-shards must be >= ${IArgs.AVAILABLE_PHYSICAL_SHARD_COUNT_RANGE.first} and <= ${IArgs.AVAILABLE_PHYSICAL_SHARD_COUNT_RANGE.last}, or -1. But current is $maxTestShards"
     )
+}
+
+private fun AndroidArgs.assertParametersConflict() {
+    if (useLegacyJUnitResult && fullJUnitResult)
+        throw FlankConfigurationError("Parameters conflict, you cannot set: `--legacy-junit-result` and `--full-junit-result` at the same time.")
 }